backend/code/services/lambdas/multi_agent_handlers/requirements.txt
boto3
PyMuPDF
python-docx
